untangle run
moved 'run plugin' logic into it's own file
