+++ run-standalone.js
@@ -1,9 +1,5 @@
 #! /usr/bin/env node
 
-var cp = require('child_process')
-var MuxRpcStream = require('muxrpc/stream')
-var toPull = require('stream-to-pull-stream')
-var pull = require('pull-stream')
 var path = require('path')
 var fs = require('fs')
 
@@ -31,39 +27,8 @@ try {
   process.exit(1)
 }
 
-function run (path, localCall) {
-  var proc = cp.spawn(path, [], {})
-  var stream = MuxRpcStream(
-    localCall,
-    require('packet-stream-codec'),
-    function () {}
-  )
+var childCall = require('./run')(pluginPath)
 
-  pull(
-    toPull.source(proc.stdout),
-    stream,
-    toPull.sink(proc.stdin)
-  )
-
-  pull(
-    toPull.source(proc.stderr),
-    toPull.sink(process.stderr)
-  )
-
-  return stream.remoteCall
-}
-
-// load and the module in the ./example directory.
-var childCall = run(
-  path.join(pluginPath, 'bin'),
-  // in practice, the localCall method is created
-  // from the local api and manifest
-  function localCall (type, name, args) {
-    console.log('CALLED', type, name, args)
-    var cb = args.pop()
-    cb(null, { okay: true })
-  }
-)
 
 var api = require('muxrpc/api')({}, manifest, childCall)
